Move drivers to separate module
